+++ crates/nwc/CHANGELOG.md
@@ -29,6 +29,11 @@
 
 - Add notification support for real-time payment updates (https://github.com/rust-nostr/nostr/pull/953)
 - Add Monitor to NostrWalletConnectOptions (https://github.com/rust-nostr/nostr/pull/989)
+- Add `NostrWalletConnectOptions::relay` method
+
+### Deprecated
+
+- Deprecate `NostrWalletConnectOptions::connection_mode` method
 
 ## v0.42.0 - 2025/05/20